feat: use newer version of wasm-tools/wit-bindgen-go
This commit switches from wasm-tools-go to the newer
go.bytecodealliance.org repository.

Right now it's pinned to the latest commit on the `main` branch
because there were some important fixes in
https://github.com/bytecodealliance/go-modules/pull/287, but once
there's been a new release we can switch to a proper version.

The `cm` package moved from a submodule to a separate module, so
the Makefile has been updated to reflect that. The `wasi-syscall`
and `wasi-cm` Make targets needed updating to work with the latest
versions of the tooling.

Unfortunately, building with this version doesn't actually work due
to an import cycle. For example, trying to build the [example] with:

```bash
$ GOTOOLCHAIN=go1.23.6 ../../go/tinygo/build/tinygo build -target=wasip2 -o add.wasm -wit-package wit/add.wit -wit-world adder main.go

package command-line-arguments
        imports runtime
        imports internal/cm
        imports encoding/json
        imports fmt
        imports os
        imports io/fs
        imports time
        imports syscall
        imports internal/cm: import cycle not allowed

package command-line-arguments
        imports runtime
        imports internal/cm
        imports encoding/json
        imports encoding/base64
        imports encoding/binary
        imports slices
        imports iter
        imports runtime: import cycle not allowed
```

I'm pretty sure this was caused by
https://github.com/bytecodealliance/go-modules/pull/279 which added the
`encoding/json` import.

[example]: https://component-model.bytecodealliance.org/language-support/go.html